Results screen and basic shop selection

+++ Assets/Scripts/Entity.cs
@@ -3,7 +3,7 @@ using System.Collections;
 
 public class Entity : MonoBehaviour 
 {
-	public int 
+	public float 
 		health,
 		maxHealth;
